+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- Added check for duplicate addresses when parsing a string.
+- Added a method to get a ByteData object to make serializing/deserializing data simpler.
+- Management of the memory segments has been moved to its own class.
+
 ## 1.0.0
 
 - Initial version.